test: test explorer opening, e2e splits for stale tests
