+++ packages/salesforcedx-vscode-apex-testing/test/playwright/specs/testExplorer.headless.spec.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { expect, type Locator } from '@playwright/test';
+import { expect } from '@playwright/test';
 
 import {
   createAndDeployApexTestClass,
@@ -20,53 +20,14 @@ import {
 
 import { test } from '../fixtures';
 import { TEST_RUN_TIMEOUT } from '../contants';
-
-// Selectors for Test Explorer UI elements
-const TEST_EXPLORER_PANEL = '[id="workbench.view.extension.test"]';
-const TEST_EXPLORER_TREE_ITEM = '[role="treeitem"]';
-const TEST_RESULTS_TAB = 'a.action-label[aria-label="Test Results"]';
-// Labels for namespace/package grouping (must match apex-testing nls)
-const LOCAL_NAMESPACE_LABEL = '(Local Namespace)';
-const UNPACKAGED_METADATA_LABEL = '(Unpackaged Metadata)';
-
-/**
- * Expands a tree row. Uses twisty click (force) then click left edge of row; one of these works for Test Explorer nodes.
- */
-const expandTreeRow = async (panel: Locator, rowLabel: string): Promise<void> => {
-  const row = panel.locator(TEST_EXPLORER_TREE_ITEM).filter({ hasText: rowLabel });
-  await row.waitFor({ state: 'visible', timeout: 15_000 });
-  const twistie = row.locator('.monaco-tl-twistie');
-
-  const isExpanded = async (): Promise<boolean> => {
-    try {
-      const collapsed = await twistie.evaluate(el => el.classList.contains('collapsed'));
-      return !collapsed;
-    } catch {
-      return false;
-    }
-  };
-
-  if (await isExpanded()) {
-    return;
-  }
-
-  try {
-    await twistie.click({ force: true });
-    await new Promise(resolve => setTimeout(resolve, 400));
-  } catch {
-    // continue
-  }
-};
-
-/** Expands Local Namespace then (Unpackaged Metadata), waiting for each child to appear before expanding the next. */
-const expandNamespaceAndPackage = async (panel: Locator): Promise<void> => {
-  await expandTreeRow(panel, LOCAL_NAMESPACE_LABEL);
-  await panel
-    .locator(TEST_EXPLORER_TREE_ITEM)
-    .filter({ hasText: UNPACKAGED_METADATA_LABEL })
-    .waitFor({ state: 'visible', timeout: 10_000 });
-  await expandTreeRow(panel, UNPACKAGED_METADATA_LABEL);
-};
+import {
+  CMD_RUN_ALL_TESTS,
+  CMD_TOGGLE_MAXIMIZED_PANEL,
+  TEST_EXPLORER_PANEL,
+  TEST_EXPLORER_TREE_ITEM,
+  TEST_RESULTS_TAB,
+  openTestExplorerAndDiscover
+} from '../helpers/testExplorerHelpers';
 
 test('Apex Tests via Test Explorer: run all, verify discovery', async ({ page }) => {
   test.setTimeout(TEST_RUN_TIMEOUT);
@@ -96,19 +57,7 @@ test('Apex Tests via Test Explorer: run all, verify discovery', async ({ page })
   });
 
   await test.step('open Test Explorer and refresh tests', async () => {
-    await executeCommandWithCommandPalette(page, 'Testing: Focus on Test Explorer View');
-    await saveScreenshot(page, 'step.explorer-focused.png');
-    const testExplorerPanel = page.locator(TEST_EXPLORER_PANEL);
-    await testExplorerPanel.waitFor({ state: 'visible', timeout: 10_000 });
-    await saveScreenshot(page, 'step.explorer-visible.png');
-    await executeCommandWithCommandPalette(page, 'Test: Refresh Tests');
-    // Discovery clears and rebuilds the tree async; wait for rebuild to settle
-    await page.waitForTimeout(1000);
-    await saveScreenshot(page, 'step.tests-refreshed.png');
-    // Wait for discovery to populate the tree (top-level "Local Namespace" node)
-    await expect(testExplorerPanel.getByText(LOCAL_NAMESPACE_LABEL)).toBeVisible({ timeout: 60_000 });
-    // Expand namespace then package so test class is visible (grouping: Namespace → Package → Class → Method)
-    await expandNamespaceAndPackage(testExplorerPanel);
+    const testExplorerPanel = await openTestExplorerAndDiscover(page);
     const testClassItem = testExplorerPanel
       .locator(TEST_EXPLORER_TREE_ITEM)
       .filter({ hasText: new RegExp(testClassName, 'i') });
@@ -117,7 +66,7 @@ test('Apex Tests via Test Explorer: run all, verify discovery', async ({ page })
   });
 
   await test.step('run all tests from Test Explorer', async () => {
-    await executeCommandWithCommandPalette(page, 'Test: Run All Tests');
+    await executeCommandWithCommandPalette(page, CMD_RUN_ALL_TESTS);
     await saveScreenshot(page, 'step.run-all-triggered.png');
   });
 
@@ -126,11 +75,13 @@ test('Apex Tests via Test Explorer: run all, verify discovery', async ({ page })
     await testResultsTab.waitFor({ state: 'visible', timeout: 30_000 });
     await testResultsTab.click();
     await saveScreenshot(page, 'step.test-results-tab.png');
-    await executeCommandWithCommandPalette(page, 'View: Toggle Maximized Panel');
+    await executeCommandWithCommandPalette(page, CMD_TOGGLE_MAXIMIZED_PANEL);
     await saveScreenshot(page, 'step.panel-maximized.png');
     await expect(page.getByText(testClassName).first()).toBeVisible({ timeout: TEST_RUN_TIMEOUT });
     await saveScreenshot(page, 'step.results-visible.png');
-    await expect(page.getByText(/passed|Passed/i)).toBeVisible({ timeout: 60_000 });
+    // Test Results panel renders "Pass Rate" / "Tests Ran" once the run completes.
+    // (Tree items have aria-label "(Passed)" but no visible "passed" text.)
+    await expect(page.getByText(/Pass Rate/i)).toBeVisible({ timeout: 60_000 });
     await saveScreenshot(page, 'step.run-done.png');
   });
